requirements.txt
openai
pymupdf 
streamlit 
python-dotenv
google.generativeai